refactor(volunteers): move age range enum in a different file
