+++ frontend/src/common/enums/age-range.enum.ts
@@ -1,0 +1,6 @@
+export enum AgeRangeEnum {
+  '0_18' = '0-18',
+  '18_30' = '18-30',
+  '30_50' = '30-50',
+  'OVER_50' = '50+',
+}
